factorizing MAMEM datasets, adding an example
