+++ moabb/datasets/__init__.py
@@ -21,6 +21,4 @@ from .mpi_mi import MunichMI
 from .ssvep_exo import SSVEPExo
 from .ssvep_nakanishi import Nakanishi2015
 from .ssvep_wang import Wang2016
-from .ssvep_mamem1 import MAMEM1
-from .ssvep_mamem2 import MAMEM2
-from .ssvep_mamem3 import MAMEM3
+from .ssvep_mamem import MAMEM1, MAMEM2, MAMEM3
